+++ raphael-solver/src/macro_solver/pareto_front/quality_pareto_front.rs
@@ -6,8 +6,8 @@ use super::{Dominate, ParetoFront};
 #[derive(Clone, Copy, PartialEq, Eq)]
 struct Value {
     cp: i16,
-    quality: u16,
-    unreliable_quality: u16,
+    quality: u32,
+    unreliable_quality: u32,
     inner_quiet: u8,
     durability: i16,
 }
@@ -37,7 +37,7 @@ impl Dominate for Value {
 
 #[derive(Clone, Copy, PartialEq, Eq, Hash)]
 struct Key {
-    progress: u16,
+    progress: u32,
     effects: Effects,
     combo: Combo,
 }
